packages/processing/requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
requests>=2.31.0
pypdf>=4.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.0.0
redis>=5.0.0
bullmq>=2.0.0
pymongo>=4.0.0
cryptography>=43.0.0
